refactor : `Price`와 `Quantity`를 공통 패키지로 이동
