+++ src/main/java/kitchenpos/common/Quantity.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.common;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
