+++ src/main/java/kitchenpos/menu/domain/MenuValidator.java
@@ -1,5 +1,6 @@
 package kitchenpos.menu.domain;
 
+import kitchenpos.common.Price;
 import kitchenpos.exception.CannotFindException;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
