+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.menu.domain.Quantity;
+import kitchenpos.common.Quantity;
 
 import javax.persistence.*;
