+++ src/test/java/kitchenpos/menu/domain/MenuValidatorTest.java
@@ -1,5 +1,7 @@
 package kitchenpos.menu.domain;
 
+import kitchenpos.common.Price;
+import kitchenpos.common.Quantity;
 import kitchenpos.exception.CannotFindException;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
